Full Functionality
The application now gives the user the option to start a new database or add a new row to the current CSV file from the current query. It is recommended that the user backup their data after each entry to mitigate the chance of accidentally erasing the current database and starting over again.